requirements.txt
Django==4.2
django-filter==23.1
djangorestframework==3.14.0
psycopg2==2.9.6
razorpay==1.3.0
